Update the png file atomically.
There seems to be a race condition between OBS Studio and this bot when
a new image is been written. The result is that sometimes OBS Studio
will show a blank image, probably failing to load a file which is at the
same time been written to.

We now write the output to a temp file and rename it to the destination
file, and the rename operation is expected to be atomic.